feat(koplugin): build static musl localsend helper binaries
Rewrite build-localsend-libs.mjs as build-localsend-bins.mjs: it now builds
the localsend-helper standalone binary (static musl for armv7/arm64, plain
cargo for macos) into apps/readest.koplugin/bin/ instead of the old FFI
cdylib into libs/. Static musl has zero glibc dependency and no
softfp/hardfloat split, so the four old glibc-pinned targets collapse into
three binary targets with the exec bit set on the copied artifact.
