+++ .gitignore
@@ -57,5 +57,5 @@ result*
 fastlane/staging/
 fastlane/metadata-play/android/en-US/changelogs
 
-apps/readest.koplugin/libs/
+apps/readest.koplugin/bin/
 apps/readest.koplugin/native/localsend-ffi/target/
